Boss #1
-first boss now operational
-more aggressively moves toward player
-shotgun attack
-toxic puddles
-arena is done for now
-boss canvas prefab made

+++ Semester2Game/Assets/Scripts/Enemies/RangedEnemy.cs
@@ -15,7 +15,7 @@ public class RangedEnemy : MonoBehaviour, IDamageable
     [SerializeField] private Transform attackPoint;
     [SerializeField] private Collider[] childHitBoxes;
 
-    [SerializeField] private float health;
+    [SerializeField] private int health;
 
     private Vector3 walkPoint;
     private bool walkPointSet;
